requirements.txt
pyautogen
python-dotenv
google-generativeai
gradio 